usage records: created by users, assigned to orgs where appropriate
