+++ apps/hash-ai-worker-ts/src/activities/create-inference-usage-record-activity.ts
@@ -2,18 +2,20 @@ import { createUsageRecord } from "@local/hash-backend-utils/service-usage";
 import type { GraphApi } from "@local/hash-graph-client";
 import type { Entity } from "@local/hash-graph-sdk/entity";
 import type { AccountId } from "@local/hash-graph-types/account";
+import type { OwnedById } from "@local/hash-graph-types/web";
 
 import type { LlmUsage } from "./shared/get-llm-response/types";
 import type { PermittedOpenAiModel } from "./shared/openai-client";
 
 export const createInferenceUsageRecordActivity = async ({
-  aiAssistantAccountId,
+  assignUsageToWebId,
   graphApiClient,
   modelName,
   usage,
   userAccountId,
 }: {
   aiAssistantAccountId: AccountId;
+  assignUsageToWebId: OwnedById;
   graphApiClient: GraphApi;
   modelName: PermittedOpenAiModel;
   usage: LlmUsage[];
@@ -30,13 +32,13 @@ export const createInferenceUsageRecordActivity = async ({
 
   return createUsageRecord(
     { graphApi: graphApiClient },
-    { actorId: aiAssistantAccountId },
     {
+      assignUsageToWebId,
       serviceName: "OpenAI",
       featureName: modelName,
-      userAccountId,
       inputUnitCount,
       outputUnitCount,
+      userAccountId,
     },
   );
 };
