+++ libs/@local/hash-backend-utils/src/service-usage.ts
@@ -17,44 +17,36 @@ import {
   systemPropertyTypes,
 } from "@local/hash-isomorphic-utils/ontology-type-ids";
 import type { AggregatedUsageRecord } from "@local/hash-isomorphic-utils/service-usage";
-import { simplifyProperties } from "@local/hash-isomorphic-utils/simplify-properties";
+import { getAggregateUsageRecords } from "@local/hash-isomorphic-utils/service-usage";
 import {
   mapGraphApiEntityToEntity,
   mapGraphApiSubgraphToSubgraph,
 } from "@local/hash-isomorphic-utils/subgraph-mapping";
-import type { ServiceFeatureProperties } from "@local/hash-isomorphic-utils/system-types/shared";
 import type { UsageRecordProperties } from "@local/hash-isomorphic-utils/system-types/usagerecord";
 import type {
   EntityRelationAndSubject,
   EntityRootType,
 } from "@local/hash-subgraph";
 import { entityIdFromComponents } from "@local/hash-subgraph";
-import {
-  getOutgoingLinkAndTargetEntities,
-  getRoots,
-} from "@local/hash-subgraph/stdlib";
-
-const generateAggregateUsageKey = ({
-  serviceName,
-  featureName,
-}: {
-  serviceName: string;
-  featureName: string;
-}) => `${serviceName}:${featureName}`;
+import { getRoots } from "@local/hash-subgraph/stdlib";
 
 /**
- * Retrieve a user's service usage
+ * Retrieve a web's service usage
  */
-export const getUserServiceUsage = async (
+export const getWebServiceUsage = async (
   context: { graphApi: GraphApi },
-  authentication: { actorId: AccountId },
   {
-    userAccountId,
     decisionTimeInterval,
-  }: { userAccountId: AccountId; decisionTimeInterval?: BoundedTimeInterval },
+    userAccountId,
+    webId,
+  }: {
+    userAccountId: AccountId;
+    decisionTimeInterval?: BoundedTimeInterval;
+    webId: OwnedById;
+  },
 ): Promise<AggregatedUsageRecord[]> => {
   const serviceUsageRecordSubgraph = await context.graphApi
-    .getEntitySubgraph(authentication.actorId, {
+    .getEntitySubgraph(userAccountId, {
       filter: {
         all: [
           generateVersionedUrlMatchingFilter(
@@ -66,7 +58,7 @@ export const getUserServiceUsage = async (
               {
                 path: ["ownedById"],
               },
-              { parameter: userAccountId },
+              { parameter: webId },
             ],
           },
         ],
@@ -94,126 +86,37 @@ export const getUserServiceUsage = async (
     .then(({ data }) => {
       return mapGraphApiSubgraphToSubgraph<EntityRootType>(
         data.subgraph,
-        authentication.actorId,
+        userAccountId,
       );
     });
 
   const serviceUsageRecords = getRoots(serviceUsageRecordSubgraph);
 
-  const aggregateUsageByServiceFeature: Record<string, AggregatedUsageRecord> =
-    {};
+  const aggregateUsageRecords = getAggregateUsageRecords({
+    decisionTimeInterval,
+    serviceUsageRecords,
+    serviceUsageRecordSubgraph,
+  });
 
-  for (const record of serviceUsageRecords) {
-    const linkedEntities = getOutgoingLinkAndTargetEntities(
-      serviceUsageRecordSubgraph,
-      record.metadata.recordId.entityId,
-    );
-
-    const serviceFeatureLinkAndEntities = linkedEntities.filter(
-      ({ linkEntity }) =>
-        linkEntity[0]!.metadata.entityTypeId ===
-        systemLinkEntityTypes.recordsUsageOf.linkEntityTypeId,
-    );
-    if (serviceFeatureLinkAndEntities.length !== 1) {
-      throw new Error(
-        `Expected exactly one service feature link for service usage record ${record.metadata.recordId.entityId}, got ${serviceFeatureLinkAndEntities.length}.`,
-      );
-    }
-
-    const serviceFeatureEntity = serviceFeatureLinkAndEntities[0]!
-      .rightEntity[0]! as Entity<ServiceFeatureProperties>;
-
-    const { featureName, serviceName, serviceUnitCost } = simplifyProperties(
-      serviceFeatureEntity.properties,
-    );
-    if (!serviceUnitCost) {
-      throw new Error("Cannot calculate usage cost without service unit cost.");
-    }
-
-    const serviceFeatureKey = generateAggregateUsageKey({
-      serviceName,
-      featureName,
-    });
-
-    const { inputUnitCount, outputUnitCount } = simplifyProperties(
-      record.properties as UsageRecordProperties,
-    );
-
-    aggregateUsageByServiceFeature[serviceFeatureKey] ??= {
-      serviceName,
-      featureName,
-      limitedToPeriod: decisionTimeInterval ?? null,
-      totalInputUnitCount: 0,
-      totalOutputUnitCount: 0,
-      totalCostInUsd: 0,
-      last24hoursTotalCostInUsd: 0,
-    };
-    const aggregateUsage = aggregateUsageByServiceFeature[serviceFeatureKey]!;
-
-    aggregateUsage.totalInputUnitCount +=
-      inputUnitCount && inputUnitCount >= 0 ? inputUnitCount : 0;
-    aggregateUsage.totalOutputUnitCount +=
-      outputUnitCount && outputUnitCount >= 0 ? outputUnitCount : 0;
-
-    const applicablePrice = serviceUnitCost.find((entry) => {
-      const { appliesUntil, appliesFrom } = simplifyProperties(entry);
-      if (
-        appliesUntil &&
-        appliesUntil <= record.metadata.provenance.createdAtTransactionTime
-      ) {
-        return false;
-      }
-      if (!appliesFrom) {
-        return false;
-      }
-      return appliesFrom <= record.metadata.provenance.createdAtTransactionTime;
-    });
-
-    if (!applicablePrice) {
-      throw new Error(
-        `No applicable price found for service feature ${serviceFeatureKey}.`,
-      );
-    }
-
-    const { inputUnitCost, outputUnitCost } =
-      simplifyProperties(applicablePrice);
-
-    const inputCost =
-      (inputUnitCount ?? 0) *
-      (inputUnitCost && inputUnitCost >= 0 ? inputUnitCost : 0);
-    const outputCost =
-      (outputUnitCount ?? 0) *
-      (outputUnitCost && outputUnitCost >= 0 ? outputUnitCost : 0);
-    const totalCost = inputCost + outputCost;
-
-    aggregateUsage.totalCostInUsd += totalCost;
-
-    const oneDayEarlier = new Date(
-      new Date().valueOf() - 24 * 60 * 60 * 1000,
-    ).toISOString();
-    if (record.metadata.provenance.createdAtTransactionTime > oneDayEarlier) {
-      aggregateUsage.last24hoursTotalCostInUsd += totalCost;
-    }
-  }
-
-  return Object.values(aggregateUsageByServiceFeature);
+  return aggregateUsageRecords;
 };
 
 export const createUsageRecord = async (
   context: { graphApi: GraphApi },
-  authentication: { actorId: AccountId },
   {
-    userAccountId,
+    assignUsageToWebId,
     serviceName,
     featureName,
     inputUnitCount,
     outputUnitCount,
+    userAccountId,
   }: {
-    userAccountId: AccountId;
+    assignUsageToWebId: OwnedById;
     serviceName: string;
     featureName: string;
     inputUnitCount?: number;
     outputUnitCount?: number;
+    userAccountId: AccountId;
   },
 ) => {
   const properties: UsageRecordProperties = {
@@ -222,6 +125,13 @@ export const createUsageRecord = async (
     "https://hash.ai/@hash/types/property-type/output-unit-count/":
       outputUnitCount,
   };
+
+  /**
+   * We want to assign usage to the web, which may be an org, but be able to identify which users
+   * incurred which usage in an org – so we create the usage record using the user that incurred it
+   * @todo what should this be for scheduled / automatically triggered flows?
+   */
+  const authentication = { actorId: userAccountId };
 
   const hashInstanceAdminGroupId = await getHashInstanceAdminAccountGroupId(
     context,
@@ -280,19 +190,15 @@ export const createUsageRecord = async (
     {
       relation: "administrator",
       subject: {
-        kind: "account",
-        subjectId: authentication.actorId,
+        kind: "accountGroup",
+        subjectId: hashInstanceAdminGroupId,
       },
     },
     {
-      relation: "viewer",
-      subject: { kind: "account", subjectId: userAccountId },
-    },
-    {
-      relation: "viewer",
+      relation: "setting",
       subject: {
-        kind: "accountGroup",
-        subjectId: hashInstanceAdminGroupId,
+        kind: "setting",
+        subjectId: "viewFromWeb",
       },
     },
   ];
@@ -300,16 +206,16 @@ export const createUsageRecord = async (
   const usageRecordEntityUuid = generateUuid() as EntityUuid;
 
   const usageRecordEntityId = entityIdFromComponents(
-    userAccountId as OwnedById,
+    assignUsageToWebId,
     usageRecordEntityUuid,
   );
 
   const createdEntities = await Entity.createMultiple(
     context.graphApi,
-    { actorId: authentication.actorId },
+    authentication,
     [
       {
-        ownedById: userAccountId as OwnedById,
+        ownedById: assignUsageToWebId,
         draft: false,
         entityUuid: usageRecordEntityUuid,
         properties,
@@ -317,7 +223,7 @@ export const createUsageRecord = async (
         relationships: entityRelationships,
       },
       {
-        ownedById: userAccountId as OwnedById,
+        ownedById: assignUsageToWebId,
         draft: false,
         properties: {},
         linkData: {
@@ -336,7 +242,7 @@ export const createUsageRecord = async (
 
   if (!usageRecordEntity) {
     throw new Error(
-      `Failed to create usage record entity for user ${userAccountId}.`,
+      `Failed to create usage record entity for webId ${assignUsageToWebId}.`,
     );
   }
